Move metadata module up a level
Move qgis specific functions to util.qgis.layer
